Add method to store and load scheduler in json
Add encoders for metrics, runners, and metrics_to_eval.
Add funcs to save and load a scheduler from json. It currently is having problems resuming a run because of the MetricToEval class encoding/decoding.
But it does let you plot it from json.

To also load from json, you will have to repass in the wrapper.
Because the wrapper was having problems with encoding and decoding to json
so just having your repass it in.