Add function to convert wpr prms to boa and func to convert boa prms to wpr
function that converts a dictionary with parameters in lots of different section into what boa needs, and prepends them with the json mapping, and then returns the mapping. normalize config can call this, and then saves the mapping. There is also a utility function to undo the mapping as well.