Add __main__.py script to run python wrapper
__main__.py can take command line arguments for python wrapper class,
wrapper dir, and other things so that you can pass in a wrapper
you wrote in python and it will handle the run script for you.
It will also save your scheduler to a json file afterwards as well.

This will work for standard run scripts. Not more exotic run scripts.