Inherit stand_alone_pkg wrapper from BaseWrapper
stand alone opt pkg did not inherit from BaseWrapper because
of pathing things for testing, because that is resolved now.
So it is simpler to implement the Wrapper as an inheritance
from BaseWrapper

This also uncovered a bug in normalize_config, which happens
in all load_jsonlike functions. If a config is passed into
__main__ that has already had its parameters normalized, then it
crashes because then parameters become a list and normalize_config
always assumes a dictionary. Now if it is already a list, it only
runs the other normalizations and doesn't normalize parameters