Add method best_fitted_trials to Scheduler
best_fitted_trials is a method that combines
get_best_trial for single obj optimizations and
get_pareto_optimal_parameters for multi obj optimizations.
It also has a backup for if the user uses a
generation step that doesn't support get_best_pareto_front
and don't include objective thresholds
but is still a multi obj optimization. It estimates
the objective thresholds and then recalcs it.