Run model script can now set trial and fetch data
Some models might be more convenient to set the trial status
in the run model script (if you are running it not on a remote
location, you are directly running the model in that
language, or if you are getting an exit code, so you don't want
to rely on logs to set the trial status.

In the same vein, it might be the most convenient to get the
trial data right then and there as well. So if you right out
the trial data in run_model, then BOA will ingest it when it
sees that the trial is complete.