Update ModularMetric to directly save metric_to_eval instead of using additional class
ModularMetric use to use a MetricToEval class for json serialization, which wasn't necessary
at all. Removed that class, and moved the functionality onto ModularMetric. This reduced
clutter and confusion. Now when you look at a metric, metric.f/metric.metric_to_eval is
the function, instead of having todo metric.f.func/metric.metric_to_eval.func Which feels
much easier and cleaner. This also made it easier to make all the built--in metrics
subclasses, which allowed easier documentation with sphinx rendering.

Added a test that saves and loads a scheduler and resumes an experiment to check that
the scheduler saving works. Added another test to save an arbitrary callable that is
passed in. it doesn't work right now, marked as skip and made an issue to get that
functionality for scheduler saving and loading with arbitrary callables working at
some point.