Update some docs in response to PR, fix macos build
PR gave feedback on missing info on passthrough metric,
on how to use the experiment_dir in the config file,
and changed the contribution guidelines to say fork
the repo instead of clone it.

Also changed the the x86mac install back to how
it was before before we switched it to nightly
pytorch since nightly pytorch is broken right now.