Add additional function to fetch all metric for a trial at once
This function allows you to just do all of your metrics in one bulk bunch and return them instead of having to return them one at a time based on the name of the metric that is calling said fetch_trial_data func. Thous the func names probably needs work.