Move to using pyproject.toml instead of setup.py
most of project metadata is now describe in pyproject.toml now
since that is where python is going to.
The dependencies are pulled from the environment.yml files dynamically and listed in the setup.py file.
Everything else is listed in the pyproject.toml file.
Also added a test in the CI script for github to test if the project builds.