Controller and __main__ can be passed scheduler.json to resume run
Controller and __main__ can be passed a scheduler.json file and they will resume experiment run if the experiment was stopped.
If some of the trials of the exp were mid run when the run was stopped, then 1 attempt will be made to compelte them
and then they are all marked failed since if the trial run was actually just ended too, it oculd lead to the exp just hanging.
Just ending it will allow Ax to start new trials in their place instead.