Moving all logging to a logger named 'boa'
This is becuase for now some people start runs in other processes and this makes the logs they write not write to the same place.
So as a work around for now, just make all things inside boa write with the same logger, and then any filehandler that is added will write to the same file.