streamline ScriptWrapper interface
Streamline ScriptWrapper interface so people working with ScriptWrapper cli
can just output one json file for the trialstatus and their data output
instead of two (one for each).
If they output an output.json file (or any other acceptable output file name)
and include the trial_status key, it will read that.
If they output only an output.json file (or any other acceptable outptu file)
and don't include the trial_status key, it will assume trial is complete
since it wrote data out. If it got to the point where it wrote data back to boa
which is the last thing it needs to do, then it is complete and boa can
count it as a passed trial.

Added tests for the streamlined version and a failing version to make sure passing
failing trail status back works too.