Add basic R BOA jupyter nb
Add a jupyter nb that runs the r_streamlined test
It uses the iPython.display.Code to display the R file
and the config file
and then it used jupyters input shebang operator for
terminal commands to run it. It then has some
assert statements to test it for sanity checks.
And it uses remove-cell or remove-input to hide
the guts that aren't documentation and are just
the dynamic rerendering parts