Wrapper serialization update
Make wrappers automatically serialize much more often and easier
Add in checks to no re-register Metrics, Runners, and Wrappers,
since that overrides previous ones. This is also an error
that raises to the user instead of a warning.
So users should have their Wrappers automatically register easier now, and get errors if it can't deserialize now.
Also add ability to pass in wrapper-name to cli to override default or config wrapper-name if needed