More robust way for wrapper to deserialize when moving computers
Move wrapper changing exp_dir from controller to `scheduler_from_json_snapshot`,
this allows changing the exp_dir right on the pre_loaded wrapper_dict,
which means it happens before reserialization.
Which is helpful for reloading the BOAConfig or anything else that needs to happen.
Also it is good for anyone not using the Controller class, and rolling their own stuff.